refactor: some refactor of dependency and classes
- Refactor some dependencies and moved some classes in their relevant module.
