+++ agent/pom.xml
@@ -64,6 +64,12 @@
             <version>4.4.13</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>com.squareup.okhttp3</groupId>
+            <artifactId>okhttp</artifactId>
+            <!--            <version>4.10.0</version>-->
+            <version>3.14.9</version>
+        </dependency>
 
     </dependencies>
