+++ integration/src/main/java/io/keploy/googleMaps/CustomHttpResponses.java
@@ -1,10 +1,8 @@
-package io.keploy.utils;
+package io.keploy.googleMaps;
 
-import okhttp3.MediaType;
 import okhttp3.Response;
 
 public class CustomHttpResponses {
     public static Response googleMapResponse;
-
     public static String googleMapResBody;
 }
