+++ integration/src/main/java/io/keploy/googleMaps/GoogleMapsInterceptor.java
@@ -10,7 +10,6 @@ import io.keploy.regression.Mock;
 import io.keploy.regression.Mode;
 import io.keploy.regression.context.Context;
 import io.keploy.regression.context.Kcontext;
-import io.keploy.utils.CustomHttpResponses;
 import io.keploy.utils.HttpStatusReasons;
 import net.bytebuddy.implementation.bind.annotation.*;
 import okhttp3.*;
