+++ keploy-sdk/pom.xml
@@ -108,7 +108,8 @@
         <dependency>
             <groupId>com.squareup.okhttp3</groupId>
             <artifactId>okhttp</artifactId>
-            <version>4.10.0</version>
+            <!--            <version>4.10.0</version>-->
+            <version>3.14.9</version>
         </dependency>
     </dependencies>
     <build>
